classification/pyproject.toml
[project]
name = "classification"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=1.4.0",
    "loguru>=0.7.3",
    "mlflow>=2.0.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "pytorch-lightning>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "timm>=0.9.0",
    "torch>=2.0.0",
    "torchmetrics>=1.0.0",
    "torchvision>=0.15.0",
]

[dependency-groups]
lint = [
    "ruff>=0.14.10",
    "ty>=0.0.7",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128"}]
torchvision = [{ index = "pytorch-cu128"}]

[tool.uv]
default-groups = []

[tool.ruff]
target-version = "py312"

# Set the maximum line length to 88.
line-length = 88

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["C", "E", "F", "I", "W", "D"]
ignore = [
    "D100",  # Missing docstring in public module
    "D400",  # First line should end with a period
    "D415"  # First line should end with a period, question mark, or exclamation point
]

extend-select = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = ["B"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ty.src]
include = ["src"]

[tool.ty.terminal]
error-on-warning = true
